# API-Gateway/src/main/resources/openapi/documentation.yaml
openapi: "3.0.3"
info:
  title: "API_Gateway API"
  description: "Api central que permite el acceso a los microservicios de reservas, salas y usuarios."
  version: "1.0.0"
servers:
- url: "http://app.iesluisvives.org:1212"
paths:
  /users:
    get:
      description: "Devuelve una lista de todos los usuarios existentes."
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDataDTO"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuarios no encontrados"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"

        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Esta operación no está permitida para los usuarios que no son administradores."
  /users/{id}:
    delete:
      description: "Elimina el usuario con el id especificado."
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "b816165e-197d-454f-97fc-d08a15648c8f"
      responses:
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                InvalidId:
                  value: "El id introducido no es válido:"
                DataNotEmpty:
                  value: "Se deben actualizar o eliminar las reservas asociadas a este usuario antes de continuar con la operación."

        "204":
          description: "No Content"
          content:
            '*/*':
              schema:
                type: "object"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Esta operación no está permitida para los usuarios que no son administradores."
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario no encontrado con el id: b816165e-197d-454f-97fc-d08a15648c8f"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
    get:
      description: "Devuelve el usuario con el id especificado."
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "b816165e-197d-454f-97fc-d08a15648c8f"
      responses:
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                InvalidId:
                  value: "El id introducido no es válido:"
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario no encontrado con el id: b816165e-197d-454f-97fc-d08a15648c8f"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Esta operación no está permitida para los usuarios que no son administradores."
    put:
      description: "Actualiza el usuario con el id especificado."
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "b816165e-197d-454f-97fc-d08a15648c8f"
      requestBody:
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/UserUpdateDTO"
        required: true
      responses:
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                InvalidId:
                  value: "El id introducido no es válido:"
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario no encontrado con el id: b816165e-197d-454f-97fc-d08a15648c8f"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Esta operación no está permitida para los usuarios que no son administradores."
  /users/active/{id}/{active}:
    put:
      description: "Actualiza el estado \"activo\" del usuario con el id especificado."
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "b816165e-197d-454f-97fc-d08a15648c8f"
      - name: "active"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "true"
      responses:
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                InvalidId:
                  value: "El id introducido no es válido:"
                InvalidBoolean:
                  value: "El estado de actividad introducido del usuario no es valido:"
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario no encontrado con el id: b816165e-197d-454f-97fc-d08a15648c8f"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Esta operación no está permitida para los usuarios que no son administradores."
  /users/credits/{id}/{creditsAmount}:
    put:
      description: "Actualiza los créditos del usuario con el id especificado."
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "b816165e-197d-454f-97fc-d08a15648c8f"
      - name: "creditsAmount"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "25"
      responses:
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                InvalidId:
                  value: "El id introducido no es válido:"
                InvalidCreditsAmount:
                  value: "El número de créditos introducido no es válido:"
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario no encontrado con el id: b816165e-197d-454f-97fc-d08a15648c8f"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Esta operación no está permitida para los usuarios que no son administradores."
  /users/login:
    post:
      description: "Loguea al usuario con el email y contraseña especificados."
      requestBody:
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/UserLoginDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserTokenDTO"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario o contraseña incorrectos."
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario o contraseña incorrectos."
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Este usuario ha sido dado de baja."
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                Example#1:
                  value: "Error interno del servidor."
                Example#2:
                  value: "EL servidor no esta disponible en este momento."
  /users/me:
    get:
      description: "Devuelve la informacion del usuario logueado."
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario no encontrado con el id: b816165e-197d-454f-97fc-d08a15648c8f"
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Este usuario ha sido dado de baja."
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
    put:
      description: "Actualiza la informacion del usuario logueado."
      requestBody:
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/UserUpdateDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Usuario no encontrado con el id: b816165e-197d-454f-97fc-d08a15648c8f"
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Este usuario ha sido dado de baja."
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
  /users/register:
    post:
      description: "Registra un nuevo usuario."
      requestBody:
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/UserRegisterDTO"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserTokenDTO"
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
                example: "El email introducido ya está en uso."
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
...
  /bookings/not-finished/user/{id}:
    get:
      description: "Devuelve todas las reservas no finalizadas de un usuario."
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            example: "fa5d82ed-fcb3-479a-97ca-334c1f46a2ff"
      responses:
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                InvalidID:
                  value: "El id debe ser un UUID válido"
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingDataDTO"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "No estas autorizado para realizar esta acción"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "No se ha encontrado el usuario con el id: fa5d82ed-fcb3-479a-97ca-334c1f46a2ff"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
...
  /bookings/user/{id}:
    get:
      description: "Devuelve todas las reservas de un usuario."
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          example: "fa5d82ed-fcb3-479a-97ca-334c1f46a2ff"
      responses:
        "400":
          description: "Bad Request"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                InvalidID:
                  value: "El id debe ser un UUID válido"
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingDataDTO"
        "401":
          description: "Unauthorized"
          content:
            '*/*':
              schema:
                type: "string"
                example: "No estas autorizado para realizar esta acción"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
                example: "No se ha encontrado el usuario con el id: fa5d82ed-fcb3-479a-97ca-334c1f46a2ff"
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
                example: "Error interno del servidor"
components:
  schemas:
    UserResponseDTO:
      type: "object"
      properties:
        uuid:
          type: "string"
          example: "fa5d82ed-fcb3-479a-97ca-334c1f46a2ff"
        name:
          type: "string"
          example: "Maria Fernanda"
        username:
          type: "string"
          example: "mariafernanda"
        email:
          type: "string"
          example: "mariafernanda@gmail.com"
        password:
          type: "string"
          example: "password12345"
        avatar:
          type: "string"
          example: "placeholder.png"
        userRole:
          type: "array"
          items:
            type: "string"
            example: "USER"
        credits:
          type: "integer"
          format: "int32"
          example: 20
    UserDataDTO:
      type: "object"
      properties:
        data:
          type: "array"
          items:
            $ref: "#/components/schemas/UserResponseDTO"
    UserUpdateDTO:
      type: "object"
      properties:
        name:
          type: "string"
          example: "Maria Fernanda"
        username:
          type: "string"
          example: "mariafernanda"
        avatar:
          type: "string"
          example: "placeholder.png"
        userRole:
          type: "array"
          items:
            type: "string"
            example: "USER"
        credits:
          type: "integer"
          format: "int32"
          example: 20
    UserLoginDTO:
      type: "object"
      properties:
        username:
          type: "string"
          example: "mariafernanda"
        password:
          type: "string"
          example: "password12345"
    UserTokenDTO:
      type: "object"
      properties:
        user:
          $ref: "#/components/schemas/UserResponseDTO"
        token:
          type: "string"
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6Ik1hcmlhIEZlcm5hbmRhIiwiaWF0IjoxNTE2MjM5MDIyfQ.iZxaYMH2xcsKWtaAhjj5gXVvvyG5MXyyRt-CAoCpp40"
    UserRegisterDTO:
      type: "object"
      properties:
        name:
          type: "string"
          example: "Maria Fernanda"
        username:
          type: "string"
          example: "mariafernanda"
        email:
          type: "string"
          example: "mariafernanda@gmail.com"
        password:
          type: "string"
          example: "password12345"
        avatar:
          type: "string"
          example: "placeholder.png"
        userRole:
          type: "array"
          items:
            type: "string"
            example: "USER"
...
    BookingResponseDTO:
      type: "object"
      properties:
        uuid:
          type: "string"
          example: "51e1e0a2-2d25-4463-a0d6-d14c865d37d3"
        userId:
          type: "string"
          example: "fa5d82ed-fcb3-479a-97ca-334c1f46a2ff"
        userName:
          type: "string"
          example: "mariafernanda"
        spaceId:
          type: "string"
          example: "c02bc82f-909d-407f-b2bd-02b00dadc3fb"
        spaceName:
          type: "string"
          example: "Departamento de FOL"
        image:
          type: "string"
          example: "placeholder.png"
        startTime:
          type: "string"
          example: "2023-06-14T:08:25.000"
        endTime:
          type: "string"
          example: "2023-06-14T:09:00.000"
        observations:
          type: "string"
          example: "Reunión de trabajo"
        status:
          type: "string"
          example: "PENDING"
    BookingDataDTO:
      type: "object"
      properties:
        data:
          type: "array"
          items:
            $ref: "#/components/schemas/BookingResponseDTO"
...